refactoring

+++ packages/app/src/gql/mutation/setZoneStatus.js
@@ -3,7 +3,7 @@ import gql from 'graphql-tag';
 import SelectPage from '../fragment/page';
 
 const template = gql`
-mutation SetZoneStatus($timestamp: String!, $id: String!, $index: Int!, $status: String!){
+mutation SetZoneStatus($timestamp: String!, $id: String!, $index: Float!, $status: String!){
     setZoneStatus(
         timestamp: $timestamp
         id: $id
